Bỏ border của bottomPanel ở BangChamCongGUI Thêm 2 bảng cho thông tin ngày nghỉ/tăng ca ở bottomPanel của BangChamCongGUI Chỉnh lại data type của soGioOT từ int sang float Thêm Dialog sửa bảng chấm công

+++ src/DAO/BangChamCongDAO.java
@@ -94,7 +94,7 @@ public class BangChamCongDAO implements DAOInterface<BangChamCongDTO>{
 			ps.setInt(4, bcc.getSoNgayLam());
 			ps.setInt(5, bcc.getSoNgayNghiPhep());
 			ps.setInt(6, bcc.getSoNgayNghiKhongPhep());
-			ps.setInt(7, bcc.getSoGioOT());
+			ps.setFloat(7, bcc.getSoGioOT());
 			ps.setString(8, bcc.getMaNV());
 			
 			result = ps.executeUpdate();
